requirements-account.txt
django-user-accounts==2.0.3
pinax-eventlog==2.0.3
pinax-webanalytics==4.0.2
certifi==2018.1.18
django-appconf==1.0.2
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django-imagekit==4.0.2
django-versatileimagefield==1.9
jsonfield==2.0.2
pilkit==2.0
Pillow==5.0.0
pinax-images==3.0.1
pipenv==11.9.0
pytz==2018.3
six==1.11.0
virtualenv==15.2.0
virtualenv-clone==0.3.0
